==> 51 Combined NSE BSE Files.ipynb <==
######### for this program to give accurate results, combined_df has to be run again if any code/name changes are there or
########## including any additions
########## companymaster has to be uptodate
######### shortlisted stocks - averages for nse/bse has to be updated
######## market cap from screener has to be updated

from platform import python_version
python_version()

### this program gives top 50 deliveries by value and volume for fno and non-fno. 
#### it also creates fno_csv and non_fno_csv which is used by multiple scrips program

### caution - if data file is already read, then it skips reading it. So if data is updated, close the program and restart it
#### to enable data file to be read again


### lenovo desktop file path
master_file_path = "C:\\Users\\User\\Dropbox\\SQL Stock Market Data\\Python Code from ChatGPT\\Master File\\CompanyMaster.csv"
directory = "C:\\Users\\User\\Dropbox\\SQL Stock Market Data\\Data Files\\"  # Replace with the path to your directory
output_path = "C:\\Users\\User\\Dropbox\\SQL Stock Market Data\\Python Code from ChatGPT\\Output File\\"
combined_file_path = "C:\\Users\\User\\Dropbox\\SQL Stock Market Data\\Delivery Price Combined Files\\"

column_order = ['NSE_Code','BSE_Code','Date','NSE_Delivery','BSE_Delivery']





#### Lenovo laptop file path

master_file_path = "C:\\Users\\sarda\\Dropbox\\SQL Stock Market Data\\Python Code from ChatGPT\\Master File\\CompanyMaster.csv"
directory = "C:\\Users\\sarda\\Dropbox\\SQL Stock Market Data\\Data Files\\"  # Replace with the path to your directory
output_path = "C:\\Users\\sarda\\Dropbox\\SQL Stock Market Data\\Python Code from ChatGPT\\Output File\\"

combined_file_path = "C:\\Users\\sarda\\Dropbox\\SQL Stock Market Data\\Delivery Price Combined Files\\"

column_order = ['NSE_Code','BSE_Code','Date','NSE_Delivery','BSE_Delivery']

import os
import csv
import pandas as pd







file_names = ['BSE Gross Deliverable.txt', 'NSE Delivery.csv', 'NSE EQ.csv', 'EQ ISIN.csv']

while True:
    user_input = input("Enter a date in the 'yymmdd' format (or press Enter to exit): ")

    # Check if the user pressed Enter with an empty input
    if not user_input:
        print("Exiting the program.")
        break

    # Check if the input has 6 characters and consists of digits
    if len(user_input) == 6 and user_input.isdigit():
        all_files_exist = True
        updated_file_names = []

        for file_name in file_names:
            full_file_name = os.path.join(directory, f"{file_name.split('.')[0]} {user_input}.{file_name.split('.')[-1]}")
            file_name_date = f"{file_name.split('.')[0]} {user_input}.{file_name.split('.')[-1]}"
            if not os.path.exists(full_file_name):
                all_files_exist = False
                break
            updated_file_names.append(file_name_date)

        if all_files_exist:
            print(f"All required files for the date '{user_input}' exist.")
            # You can use the updated_file_names list for further processing
            print("Updated file names:", updated_file_names)
            break  # Exit the loop if all files exist
        else:
            print("Not all required files exist. Please check the files.")
    else:
        print("Invalid input. Please enter a date in the 'yymmdd' format (e.g., 220101).")
        
        



nse_dely_file = [updated_file_names[1]]
nse_price_file = [updated_file_names[2]]
bse_dely_file = [updated_file_names[0]]
bse_price_file = [updated_file_names[3]]
print(nse_dely_file,nse_price_file,bse_dely_file,bse_price_file)

def nse_data(nse_file_name):
    
    
    
    nse_df = pd.DataFrame(columns=['NSE_Code','Date','NSE_Volume','NSE_Delivery','Segment'])
    dely_segment = ['EQ','BE','SM','ST']
    
    for file_name in nse_file_name:
        file_path = os.path.join(directory, file_name)
        date = file_name[13:19]
        
        with open(file_path,'r') as file:
            reader = csv.reader(file)
            for i, line in enumerate(reader, start=1):
                if i >= 5:
                    line_str = ','.join(line)
                    data = line_str.split(',')
                    if data[3] in dely_segment:
                        nse_code = data[2]
                        dely = data[5]
                        segment = data[3]
                        volume = data[4]
                        df_to_append = pd.DataFrame({'NSE_Code': [nse_code], 'Date': [date], 'NSE_Volume':volume,
                                                     'NSE_Delivery': [dely], 'Segment': [segment]})
                        nse_df = pd.concat([nse_df, df_to_append], ignore_index=True)
                        #nse_df = nse_df.append({'NSE_Code': nse_code, 'Date': date, 'NSE_Delivery':dely, 'Segment':segment}, ignore_index=True)
    
    return nse_df
                
 
def bse_data(bse_file_name):
    
    
    
    # Create an empty DataFrame to store the results
    bse_df = pd.DataFrame(columns=['BSE_Code', 'Date','BSE_Delivery'])
    dely_segment = ['EQ','BE','SM','ST']

    

    for file_name in bse_file_name:
        file_path = os.path.join(directory, file_name)
        date = file_name[22:28] 
    
        with open(file_path, 'r') as file:
            reader = csv.reader(file)
            for i, line in enumerate(reader, start=1):
                if i >= 2:  # Start reading from line 5
                    line_str = ','.join(line)
                    data = line_str.split('|')
                    dely = data[2]
                    bse_code = data[1]
                    df_to_append = pd.DataFrame({'BSE_Code': [bse_code], 'Date': [date], 'BSE_Delivery':[dely]}) 
                    bse_df = pd.concat([bse_df,df_to_append], ignore_index=True)
                        
    return bse_df





def nse_price(nse_price_file):
    nse_df = pd.DataFrame(columns=['NSE_Code', 'Date', 'Price', 'PrevClose', 'Segment', 'Traded'])
    dely_segment = ['EQ', 'BE', 'SM', 'ST']

    for file_name in nse_price_file:
        file_path = os.path.join(directory, file_name)
        date = file_name[7:13]

        with open(file_path, 'r') as file:
            reader = csv.reader(file)
            for row in reader:
                if row and row[1] in dely_segment:  # Check if the row is not empty and segment matches
                    nse_code = row[0]
                    price = row[5]
                    prev_close = row[7]
                    segment = row[1]
                    traded = row[8]
                    df_to_append = pd.DataFrame({'NSE_Code': [nse_code], 'Date': [date], 'Price': [price], 
                                                 'PrevClose': [prev_close],'Segment': [segment], 'Traded': [traded]})
                    nse_df = pd.concat([nse_df, df_to_append], ignore_index=True)

    return nse_df

def bse_price(bse_price_file):
    bse_df = pd.DataFrame(columns=['BSE_Code', 'Date', 'Price', 'PrevClose', 'BSE_Volume','Segment'])
    dely_segment = ['EQ', 'BE', 'SM', 'ST']

    for file_name in bse_price_file:
        file_path = os.path.join(directory, file_name)
        date = file_name[8:14]

        with open(file_path, 'r') as file:
            reader = csv.reader(file)
            for row in reader:
                #if row and row[1] in dely_segment:  # Check if the row is not empty and segment matches
                    bse_code = row[0]
                    price = row[7]
                    prev_close = row[9]
                    bse_group = row[2]
                    volume = row[11]
                    df_to_append = pd.DataFrame({'BSE_Code': [bse_code], 'Date': [date], 'Price': [price], 
                                                 'PrevClose': [prev_close],'BSE_Volume':volume ,'Segment': [bse_group]})
                    bse_df = pd.concat([bse_df, df_to_append], ignore_index=True)

    # Drop the first row, which is a header row, and reset the index
    bse_df = bse_df.iloc[1:].reset_index(drop=True)

    return bse_df


def add_nse_prefix(df, column_name):
    # Create a copy of the DataFrame to avoid SettingWithCopyWarning
    df = df.copy()

    # Add "NSE:" prefix to the specified column
    df.loc[:, column_name] = "NSE:" + df[column_name].astype(str)

    return df

### getting unique symbols from fno
nse_fno_file = sorted([file for file in os.listdir(directory) if file.startswith('NSE FNO')],reverse=True)[:1]

for file_name in nse_fno_file:
    file_path = os.path.join(directory, file_name)
    fno_df = pd.read_csv(file_path)
    fno_df = fno_df.drop_duplicates(subset ='SYMBOL')
    fno_symbols = fno_df['SYMBOL']
    fno_symbols = fno_symbols.reset_index(drop=True)

print('processing nse_dely')
nse_dely_data = nse_data(nse_dely_file)

print('processing bse_dely')
bse_dely_data = bse_data(bse_dely_file)

print('processing nse_price')
nse_price_data = nse_price(nse_price_file)

print('processing bse_price')
bse_price_data = bse_price(bse_price_file)

print('all done, proceed')



nse_dely_data['NSE_Delivery'] = nse_dely_data['NSE_Delivery'].astype(int)
nse_dely_data['NSE_Volume'] = nse_dely_data['NSE_Volume'].astype(int)
nse_price_data['Price'] = nse_price_data['Price'].astype(float)
nse_price_data['PrevClose'] = nse_price_data['PrevClose'].astype(float)
nse_price_data['Traded'] = nse_price_data['Traded'].astype(int)

bse_dely_data['BSE_Delivery'] = bse_dely_data['BSE_Delivery'].astype(int)
bse_price_data['Price'] = bse_price_data['Price'].astype(float)
bse_price_data['PrevClose'] = bse_price_data['PrevClose'].astype(float)
bse_price_data['BSE_Volume'] = bse_price_data['BSE_Volume'].astype(int)


#nse_dely_data.info()
#bse_dely_data.info()
#nse_price_data.info()
#bse_price_data.info()

### nse_dely_data has NSE_Code, Date, NSE_Delivery, Segment
### bse_dely_data has BSE_Code, Date, BSE_Delivery

### nse_price_data has NSE_Code	Date	Price	PrevClose	Segment	Traded
### bse_price_data has BSE_Code	Date	Price	PrevClose	Segment

## this section processes nse delivery and price file. Makes sure that stocks which are in BE segment and 
## therefore dont appear in nse delivery are taken from price file and new nse merged df is generated

### merging nse price and delivery files
nse_merged_df_out = pd.merge(nse_dely_data,nse_price_data, on = 'NSE_Code', how = 'outer')

### processing merged file
# Replace 'NSE_Delivery' with 'Traded' where 'NSE_Delivery' is NaN
nse_merged_df_out['NSE_Delivery'] = nse_merged_df_out['NSE_Delivery'].fillna(nse_merged_df_out['Traded'])

# Replace nse volume with nse delivery where nse volume is NaN
nse_merged_df_out['NSE_Volume'] = nse_merged_df_out['NSE_Volume'].fillna(nse_merged_df_out['NSE_Delivery'])

# Replace 'Date_x' with 'Date_y' where 'Date_x' is NaN
nse_merged_df_out['Date_x'] = nse_merged_df_out['Date_x'].fillna(nse_merged_df_out['Date_y'])

# Replace 'Segment_x' with 'Segment_y' where 'Segment_x' is NaN
nse_merged_df_out['Segment_x'] = nse_merged_df_out['Segment_x'].fillna(nse_merged_df_out['Segment_y'])

# Rename 'Segment_x' to 'Segment'
nse_merged_df_out.rename(columns={'Segment_x': 'Segment'}, inplace=True)

# Rename 'Date_x' to 'Date'
nse_merged_df_out.rename(columns={'Date_x': 'Date'}, inplace=True)

# Drop the 'Date_y' column
nse_merged_df_out.drop(columns=['Date_y'], inplace=True)

# Drop the 'Segment_y' column
nse_merged_df_out.drop(columns=['Segment_y'], inplace=True)

# Delete the 'Traded' column after replacement
nse_merged_df_out.drop(columns=['Traded'], inplace=True)

# Reset the index
nse_merged_df_out.reset_index(drop=True, inplace=True)

# add traded_value column

nse_merged_df_out['Traded_Value_lacs'] = (nse_merged_df_out['NSE_Delivery'] * nse_merged_df_out['Price'] / 100000).round(2)

#print(nse_merged_df_out)

nse_col_order = ['Date','NSE_Code','NSE_Volume','NSE_Delivery','Price','PrevClose','Traded_Value_lacs','Segment']
nse_merged_df_out = nse_merged_df_out[nse_col_order]

#### nse_merged_df_out is NSE_Code	Date	NSE_Delivery	Segment	Price	PrevClose	Traded_Value_lacs


###merging bse price and delivery files

bse_merged_df_out = pd.merge(bse_dely_data,bse_price_data, on = 'BSE_Code', how = 'outer')



### processing bse merged file

bse_merged_df_out['Traded_Value_lacs'] = (bse_merged_df_out['BSE_Delivery'] * bse_merged_df_out['Price'] / 100000).round(2)

# Rename 'Date_x' to 'Date'
bse_merged_df_out.rename(columns={'Date_x': 'Date'}, inplace=True)

# Drop the 'Date_y' column
bse_merged_df_out.drop(columns=['Date_y'], inplace=True)
bse_merged_df_out = bse_merged_df_out[bse_merged_df_out['BSE_Code'] <= '599999']

bse_merged_df_out.reset_index(drop=True, inplace = True)

#print(bse_merged_df_out)

bse_col_order = ['Date','BSE_Code','BSE_Volume','BSE_Delivery','Price','PrevClose','Traded_Value_lacs','Segment']
bse_merged_df_out = bse_merged_df_out[bse_col_order]

#### bse_merged_df_out is BSE_Code	Date	BSE_Delivery	Price	PrevClose	Segment	Traded_Value_lacs

### saving nse and bse files separately

nse_file = combined_file_path + 'nse_' + user_input + '.csv'
bse_file = combined_file_path + 'bse_' + user_input + '.csv'

#df.to_csv('sample.csv', index=False)

nse_merged_df_out.to_csv(nse_file, index = False)
bse_merged_df_out.to_csv(bse_file, index = False)



#nse_merged_df_out.info()
#nse_merged_df_out

#bse_merged_df_out.info()
#bse_merged_df_out

## reading master file
### master file (csv) is Company Name	CD_NSE Symbol	CD_BSE Code	CD_Is Subsidiary	CD_Sector	CD_Industry1

#### TO FIX - MASTER FILE DOES NOT HAVE SM SEGMENT CODES

master_file = pd.read_csv(master_file_path, dtype={'CD_BSE Code': str})


###df.rename(columns={'old_column_name': 'new_column_name'}, inplace=True)
##df.rename(columns={'old_column_name': 'new_column_name'}, inplace=True)
### Inplace method (modifies the original DataFrame)
###df.drop(columns=['column1', 'column2'], inplace=True)

master_file.drop(columns=['CD_Is Subsidiary','CD_Sector','CD_Industry1'], inplace = True )
master_file.rename(columns={'CD_NSE Symbol': 'NSE_Code', 'CD_BSE Code': 'BSE_Code'}, inplace= True )


#### master_file df is Company Name	NSE_Code	BSE_Code

#print(master_file)

## Merge NSE df with master file 

master_file = pd.merge(master_file,nse_merged_df_out, on= "NSE_Code", how = 'outer')
#master_file.dropna(subset=['Company Name'], inplace=True)


## merge bse df  with master file 
master_file = pd.merge(master_file,bse_merged_df_out, on ="BSE_Code", how = "outer")
#master_file.dropna(subset=['Company Name'], inplace=True)

# Drop records where both 'NSE_Code' and 'BSE_Code' are NaN or blank
master_file.dropna(subset=['NSE_Code', 'BSE_Code'], how='all', inplace=True)

# Reset the index after dropping rows
master_file.reset_index(drop=True, inplace=True)

# Drop records where both 'NSE_Delivery' and 'BSE_Delivery' are 0 or blank or NaN
master_file = master_file[~((master_file['NSE_Delivery'] == 0) & (master_file['BSE_Delivery'] == 0))]

# Reset the index after dropping rows
master_file.reset_index(drop=True, inplace=True)


# Trim "Company Name" column to a maximum length of 20 characters
master_file['Company Name'] = master_file['Company Name'].str.slice(0, 30)


# Reset the index after dropping rows
master_file.reset_index(drop=True, inplace=True)

#print(master_file)


# Identify the data types of each column
column_data_types = master_file.dtypes

# Replace NaN values in object columns with empty strings ("")
object_columns = master_file.select_dtypes(include='object').columns
master_file[object_columns] = master_file[object_columns].fillna("")

# Replace NaN values in float/integer columns with 0
numeric_columns = master_file.select_dtypes(include=['float64', 'int64']).columns
master_file[numeric_columns] = master_file[numeric_columns].fillna(0)

# Print the resulting DataFrame
#print(master_file)

#master_file.info()

# Define a custom function to replace Price_x with Price_y if Price_x is 0
def replace_price_x(row):
    if row['Price_x'] == 0:
        return row['Price_y']
    else:
        return row['Price_x']

    
def replace_prevclose_x(row):
    if row['PrevClose_x'] == 0:
        return row['PrevClose_y']
    else:
        return row['PrevClose_x']
    
    
def replace_segment(row):
    if row['Segment_x'] == "":
        return row['Segment_y']
    else:
        return row['Segment_x']
    
def replace_date(row):
    if row['Date_x'] == "":
        return row['Date_y']
    else:
        return row['Date_x']

# Apply the custom function to the DataFrame
master_file['Price_x'] = master_file.apply(replace_price_x, axis=1)
master_file['PrevClose_x'] = master_file.apply(replace_prevclose_x, axis=1)

# Rename 'Price_x' column to 'Price'
master_file.rename(columns={'Price_x': 'Price'}, inplace=True)
master_file.rename(columns={'PrevClose_x':'PrevClose'}, inplace=True)

### segment

master_file['Segment_x'] = master_file.apply(replace_segment, axis = 1)
master_file.rename(columns={'Segment_x' :'Segment'}, inplace = True)

#Date
master_file['Date_x'] = master_file.apply(replace_date, axis = 1)
master_file.rename(columns={'Date_x' :'Date'}, inplace = True)

### add total delivery column
master_file['Total_Delivery'] = master_file['NSE_Delivery'] + master_file['BSE_Delivery']
master_file['Total_Delivery'] = master_file['Total_Delivery'].astype(int)
master_file = master_file[master_file['Total_Delivery'] >= 1]

### add total volume column
master_file['Total_Volume'] = master_file['NSE_Volume'] + master_file['BSE_Volume']
master_file['Total_Volume'] = master_file['Total_Volume'].astype(int)


## add total traded value colum

master_file['Total_Traded_Value'] = ( master_file['Total_Delivery'] * master_file['Price'] ) / 100000



### reorder columns

columns_to_copy = ['Date','Company Name','Total_Volume','Total_Delivery', 'Price','PrevClose','Total_Traded_Value' ,'NSE_Code','BSE_Code','Segment']

# Create a new DataFrame 'final_master_file' with the desired columns
#final_master_file = master_file[columns_to_copy].copy()



#master_file


master_file_columns = ['Date','Company Name','NSE_Code','BSE_Code','Total_Volume','Total_Delivery', 'Price','PrevClose',
                       'Total_Traded_Value' , 'NSE_Volume','NSE_Delivery','BSE_Volume','BSE_Delivery','Segment']
final_master_file = master_file[master_file_columns]
master_file_csv = combined_file_path + 'combined_' + user_input + '.csv'

#df.to_csv('sample.csv', index=False)

final_master_file.to_csv(master_file_csv, index = False)


#### create sme file
sme_segment = ['SM','ST','M ','MT']
sme_delivery = final_master_file[final_master_file['Segment'].isin(sme_segment)]
sme_delivery.loc[sme_delivery['NSE_Code'] == '', 'NSE_Code'] = sme_delivery['BSE_Code']
sme_delivery.reset_index(drop=True, inplace=True)

# Drop records where Total_Traded_Value is below 1000
final_master_file = final_master_file[final_master_file['Total_Traded_Value'] >= 1000]


###creating separate file for fno delivery and non fno delivery

# Filter final_master_file to get fno_delivery by checking if NSE_Code is in fno_symbols
fno_delivery = final_master_file[final_master_file['NSE_Code'].isin(fno_symbols)]
fno_delivery_final = fno_delivery[['NSE_Code', 'NSE_Volume','Total_Delivery', 'Price','PrevClose', 
                                   'Total_Traded_Value']].copy()


# Create non_fno_delivery by dropping rows where NSE_Code is in fno_symbols
non_fno_delivery = final_master_file[~final_master_file['NSE_Code'].isin(fno_symbols)]

# Reset index inplace for fno_delivery and non_fno_delivery
fno_delivery_final.reset_index(drop=True, inplace=True)
non_fno_delivery.reset_index(drop=True, inplace=True)

# Replace empty strings in NSE_Code with BSE_Code
#non_fno_delivery.loc[non_fno_delivery['NSE_Code'] == '', 'NSE_Code'] = non_fno_delivery['BSE_Code']


# Replace empty strings in NSE_Code with BSE_Code using a copy
non_fno_delivery_copy = non_fno_delivery.copy()
non_fno_delivery_copy.loc[non_fno_delivery_copy['NSE_Code'] == '', 'NSE_Code'] = non_fno_delivery_copy['BSE_Code']





# Replace empty strings in NSE_Code with BSE_Code
#non_fno_delivery['NSE_Code'] = non_fno_delivery.apply(lambda row: row['BSE_Code'] if row['NSE_Code'] == "" else row['NSE_Code'], axis=1)

#non_fno_delivery.loc[non_fno_delivery['NSE_Code'] == '', 'NSE_Code'] = non_fno_delivery['BSE_Code']


#sme_delivery
#non_fno_delivery

segment_var = ['SM','ST','M ']
#final_master_file.info()
 ##Count the records in each segment
segment_counts = master_file['Segment'].value_counts()

#print(segment_counts)

nse_dely_date = user_input
### creating top 50 by value and by volume for fno and non-fno

def top_n(df,n, column_name, trade_amt):
    
    df_filtered = df[df['Total_Traded_Value'] > trade_amt]
    top_n = df_filtered.nlargest(n,column_name).sort_values(column_name, ascending = False)
    return top_n

df_value_50 = top_n(non_fno_delivery,50,'Total_Traded_Value', 250)
df_volume_50 = top_n(non_fno_delivery,50,'Total_Delivery', 250)

fno_value_50 = top_n(fno_delivery,50,'Total_Traded_Value', 1000)
fno_volume_50 = top_n(fno_delivery,50,'Total_Delivery', 1000)

fno_columns = ['NSE_Code', 'Total_Delivery','Price','PrevClose','Total_Traded_Value']

fno_value_50 = fno_value_50[fno_columns]
fno_volume_50 = fno_volume_50[fno_columns]

sme_value_50 = top_n(sme_delivery,50,'Total_Traded_Value', 50)

###########################

###########################

## write to excel -- non_fno value and volume is merged by concat and saved for multiple scrips report (7-day history)
#### sme_csv is similarly saved for 7-day history report

df_non_fno = pd.concat([df_value_50, df_volume_50])
df_non_fno_wl = df_non_fno[['NSE_Code']]

df_fno = pd.concat([fno_value_50, fno_volume_50])




# Reset the index of the series
df_non_fno_wl.reset_index(drop=True, inplace=True)

#add NSE: prefix to NSE_Code
df_non_fno_wl_nse = add_nse_prefix(df_non_fno_wl,'NSE_Code')


### define csv files
output_non_fno = output_path + "0_non_fno.csv"
output_non_fno_date = output_path + nse_dely_date + "_0_non_fno_" + ".csv"
output_non_fno_wl = output_path + '_' + nse_dely_date +"_0_non_fno_wl" + ".csv"


output_fno = output_path + "fno.csv"

output_sme_nodate = output_path + 'sme_csv.csv'
output_sme = output_path + nse_dely_date + '_0_sme_' + '.csv'

### write to csv
df_non_fno.to_csv(output_non_fno, index = False)
df_non_fno.to_csv(output_non_fno_date, index = False)
df_non_fno_wl_nse.to_csv(output_non_fno_wl, index = False)

df_fno.to_csv(output_fno, index = False)

sme_delivery.to_csv(output_sme, index=False)
sme_value_50.to_csv(output_sme_nodate)






def print_formatted_rows(df, columns_to_select):
    # Iterate over the DataFrame and print each row
    for i, (_, row) in enumerate(df.iterrows(), 1):
        p_coname = row[columns_to_select[0]].ljust(40)
        p_dely = f"{row[columns_to_select[1]]:,.0f}".rjust(15)
        p_price = f"{row[columns_to_select[2]]:,.2f}".rjust(12)
        p_prevclose = f"{row[columns_to_select[3]]:,.2f}".rjust(12)
        p_trade_value = f"{row[columns_to_select[4]]:,.2f}".rjust(18)
        p_nse_code = row[columns_to_select[5]].ljust(10)
        
        print(f"{p_coname}{p_dely}{p_price}{p_prevclose}{p_trade_value}    {p_nse_code}")

        # Check if the current row index is a multiple of 5, and insert a blank line
        if i % 5 == 0:
            print()

def fno_formatted_rows(df, columns_to_select):
    
    # Iterate over the DataFrame and print each row
    for i, (_, row) in enumerate(df.iterrows(),1):
        p_nse_code = row[columns_to_select[0]].ljust(10)
        p_dely = f"{row[columns_to_select[1]]:,.0f}".rjust(15)
        p_price = f"{row[columns_to_select[2]]:,.2f}".rjust(12)
        p_prevclose = f"{row[columns_to_select[3]]:,.2f}".rjust(12)
        p_trade_value = f"{row[columns_to_select[4]]:,.2f}".rjust(12)
        
        print(f"{p_nse_code}{p_dely}{p_price}{p_prevclose}{p_trade_value}")
        
        # Check if the current row index is a multiple of 5, and insert a blank line
        if i % 5 == 0:
            print()






columns_to_select = ['Company Name', 'Total_Delivery', 'Price', 'PrevClose','Total_Traded_Value', 'NSE_Code']
col_length = [40,15,12,12,18,10]
formatted_columns = [col + ' ' * (length - len(col)) for col, length in zip(columns_to_select, col_length)]
print(formatted_columns)


##columns selection for printing
columns_to_select = ['Company Name', 'Total_Delivery', 'Price','PrevClose', 'Total_Traded_Value', 'NSE_Code']
col_length = [40,15,12,18,10]
fno_columns_to_select = ['NSE_Code', 'Total_Delivery', 'Price','PrevClose', 'Total_Traded_Value']



print("Top 50 deliveries by value/volume non-FnO FnO and SME ", nse_dely_date)
print("")
print("")
#print(nse_dely_date)
#print()
print("                          Top 50 delivery by value NON FNO ", nse_dely_date)
print("")
print("Company Name                                ", '  Dely_Qty','      Price','   PrevClose','           Value','   NSECode')
print("")
print_formatted_rows(df_value_50,columns_to_select)
print("")
print("")
print("")


print("                        Top 50 delivery by volume NON FNO ", nse_dely_date)
print("")
print("Company Name                                ", '  Dely_Qty','      Price','   PrevClose','           Value','   NSECode')
print("")

print_formatted_rows(df_volume_50,columns_to_select)
print("")
print("")
print("")


#### printing for fno
print("          Top 50 FNO Delivery by value ", nse_dely_date)
print("")
print("NSECode       ",'  Dely_Qty','      Price','   PrevClose','    Value' )
print("")
fno_formatted_rows(fno_value_50,fno_columns_to_select)
print("")
print("")
print("")

print("             Top 50 FNO Delivery by volume ", nse_dely_date)
print("")
print("NSECode       ",'  Dely_Qty','      Price','   PrevClose','   Value' )
print("")

fno_formatted_rows(fno_volume_50,fno_columns_to_select)
print("")
print("")
print("")


print("                          Top 50 delivery by value SME ", nse_dely_date)
print("")
print("Company Name                                ", '  Dely_Qty','      Price','   PrevClose','           Value','   NSECode')
print("")
print_formatted_rows(sme_value_50,columns_to_select)
print("")
print("")
print("")




sorted_df = bse_dely_data.sort_values(by='BSE_Delivery', ascending = False)
sorted_df


######### fromatting text
text = "Hello"
width = 10

left_justified_text = text.ljust(width)
print(left_justified_text)  # Output: "Hello     "

########### formatting numbers

# Example decimal value
value = 1234567.89

# Format the decimal value with 2 decimal points and commas
formatted_value = f'{value:,.2f}'

# Specify the desired width for right-justification
width = 15

# Right-justify the formatted value to the specified width
right_justified_value = formatted_value.rjust(width)

# Print the result
print(right_justified_value)

################# adding both

# Example string and number
text = "Total Traded Value:"
value = 1234567.89

# Format the number with 2 decimal points and commas
formatted_value = f'{value:,.2f}'

# Concatenate the string and the formatted number
result_string = f'{text} {formatted_value}'

# Print the result
print(result_string)


##### adding spaces
# Calculate the number of spaces needed to make the total length 50 characters
num_spaces = 30 - len(text)

# Concatenate the text with the required number of spaces and the formatted value
result_string = text + ' ' * num_spaces + formatted_value

# Print the result
print(result_string)


# Value to match against
#target_name = 'Charlie'

# Locate the record(s) matching the value in the 'Name' column
#matching_records = df[df['Name'] == target_name]

#volume = 3551452
#find_nse_symbol = nse_merged_df_out[nse_merged_df_out['NSE_Delivery']== volume]
#find_nse_symbol

import pandas as pd

# Example DataFrame (replace with your actual DataFrame)
data = {
    'Name': ['John', 'Alice', 'Bob'],
    'Age': [30, 25, 22],
    'City': ['New York', 'London', 'Paris']
}

df = pd.DataFrame(data)

# Add the 'Salary' column with the given values
df['Salary'] = [30000, 49304.35, 2938]

# Format the 'Salary' column with 2 decimal points and commas
df['Formatted Salary'] = df['Salary'].apply(lambda x: f'{x:,.2f}')

# Remove headers and iterate through rows using iterrows()
for _, row in df.iterrows():
    name = row['Name'].ljust(20)
    age = str(row['Age']).rjust(9)
    city = row['City'].ljust(30)
    salary = f"{row['Salary']:,.2f}".rjust(15)
    print(f"{name}{age}  {city}{salary}")


m_segment = final_master_file[final_master_file['Segment'] == "SM"]
m_segment

sme_value_50
